+++ internal/xsql/conn.go
@@ -60,7 +60,11 @@ func (c *Conn) BeginTx(ctx context.Context, opts driver.TxOptions) (_ driver.Tx,
 		stack.FunctionID("database/sql.(*Conn).BeginTx", stack.Package("database/sql")),
 	)
 	defer func() {
-		onDone(c.currentTx, finalErr)
+		if c.currentTx != nil {
+			onDone(c.currentTx, finalErr)
+		} else {
+			onDone(nil, finalErr)
+		}
 	}()
 
 	if c.currentTx != nil {
@@ -102,7 +106,11 @@ func (c *Conn) Begin() (_ driver.Tx, finalErr error) {
 		stack.FunctionID("database/sql.(*Conn).Begin", stack.Package("database/sql")),
 	)
 	defer func() {
-		onDone(c.currentTx, finalErr)
+		if c.currentTx != nil {
+			onDone(c.currentTx, finalErr)
+		} else {
+			onDone(nil, finalErr)
+		}
 	}()
 
 	if c.currentTx != nil {